Extracted common functionality of ContentInputElement and Interactive into EventHandlers.
