+++ src/Avalonia.Controls/Primitives/SelectingItemsControl.cs
@@ -745,7 +745,7 @@ namespace Avalonia.Controls.Primitives
                 Mark(i, false);
             }
 
-            var route = BuildEventRoute(SelectionChangedEvent);
+            var route = this.BuildEventRoute(SelectionChangedEvent);
 
             if (route.HasHandlers)
             {
